New galaxy composition check methods
Both type of Subhalos (i.e., central and satellite) are meant to meet
certain criteria in terms of the types of galaxies that are allowed to
be contained within. Some of these checks were being performed in some
parts of the code, but not only there was some code duplication here and
there, but also checks were not being ran consistently.

This commit centralises the logic of checking that the galaxy
composition of both central and satellite subhalos meets the expected
conditions, throwing an exception if a particular situation is not met.
There are three different methods to performs checks on whatever type a
given subhalo happens to be, and to performs checks with a particular
type of subhalo in mind (therefore additionally checking that the
subhalo is of the expected type). We use these new checks throughout the
code in the place where we previously had checks, and in other new
places where we think it makes sense to check these constrains.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>