Correcting how main galaxies change their type
There were a few flaws spotted by the new galaxy composition functions
indicating that main galaxies (i.e., central galaxies on central subhalos
and type 1 galaxies on satellite subhalos) were not being correctly
transferred to their descendant subhalos. The error was in the type
there were being assigned, which did not follow the expectations
subhalos are supposed to meet.

This commit fixes not only this, but also makes sure we fully skip
subhalos that are known to finish during this snapshot.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>