Reworking transfer_galaxies_to_next_snapshot
Now that most of the logic in this function has been simplified and
centralised, we can write a single loop to carry out all the logic. The
code reads much easier this way, and we have adjusted it to have all
logical pieces closely together (e.g., all galaxy transfer bits occur
within 5 lines, all mass transfer happens together).

I took the chance to move the doxygen doc of the function to be above
its definition, and removed the unnecessary Cosmology parameter.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>