Correcting subhalo transfer across snapshots
An important missing case during this transfer of galxies was that of central
subhalos that were about to disappear. These subhalos *need* to transfer their
galaxies into their descendants, unlike satellites which do not (because they
merge onto the central subhalo in merging_subhalos.cpp, transferring all their
galaxies over).

This also means that a descendant subhalo can receive galaxies from more than
one parent subhalo. As a consequence of this, the checks of the descendants
being empty needed to be done before any galaxy transfer; likewise the check on
the galaxy composition of the descendant subhalos needs to be done after all
transfers.