Gas_cooling now allows for environmental processes
- Previously, the gas_cooling function assumed that satellite subhalos
are instantaneously stripped of their hot gas. This would have made it
difficult in the near future to include more complex models. So the new
changes included:

1) call the new function environment that will internally calculate
the the stripping of hot gas from the satellite subhalos.

2) this means that the routine of gas_cooling had to be adapted to
allow for possible gas cooling to happen in satellite subhalos, that would
later feed their type=1 galaxy.

3) the code now returns a cooling rate =0 only in case where the galaxy
is a type=2, and when the subhalo hosting the galaxy has a mass of hot gas
=0.