Fixing incorrect usage of std::shared_ptr
We were creating a new shared_ptr over the address of a Galaxy object which was
already owned by a few shared_ptrs. This meant meant that when the new
std::shared_ptr was destroyed the Galaxy underneath was also destroyed,
invalidating the other shared_ptrs, and bringing chaos to shark.

We were using this new shared_ptr only to avoid changing all the rest of the
code that used the -> syntax to access the members of the Galaxy object. The
correct way of doing this, avoiding deleting the underlying object, is of
course to use a raw Galaxy pointer.

Signed-off-by: Claudia Lagos <claudia.lagos@icrar.org>