Bulge mass due to accreted stars is now followed
- Previously, the code was following only the mass that was formed via
starbursts due to disk instabilities and mergers. That mass is by
definition in the bulge. However, another growth channel for the bulge
is the accretion of stars that were formed in the disk of the primary
and/or secondary galaxy that because of the merger is transferred to the
bulge. This is now being followed. 

- For that it was necessary to add in the components.h a property to
track that growth channel, and galaxy_mergers was edited to explicitely
track that growth.