Shark using data files shipped with the code
We now have static data files that are shipped with the code, and
installed together with the rest of the executables. On the other hand
we also have the ability to load these files from within the code. This
commit finally makes use of this ability and lift of the responsibility
from the user to specify paths to these data files.

The gas cooling already abstracted a bit the filenames by using the
value of the lambdamodel option to determine the file, but still
required the directory where the files resided as an input. That is not
needed anymore, as it is calculated internally instead.

On the other hand, the cosmology needed the precise path to the power
spectrum file to be used. Since now we ship them inside shark, we
require the user instead to specify *which* power spectrum to use from
an enumerated list of possibilities, which we internally translate to
the corresponding file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>