Fixes for compatibility with matplotlib 1.3.x
A few minor problems were found when running the standard plots against
matplotlib 1.3.x. This version is from 2013 and is still used in some
Linux distros, so we should probably try to support it.

The following changes were required:
 * The 'viridis' colormap appeared only in mpl 1.5
 * 'linestyle'/'ls' values cannot be None
 * 'markersize' values must be numeric

I also took the opportunity to overhaul the density_contour function to
require its ax argument instead of making it optional (because all
callers were giving one).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>