Adding more documentation on shark configuration
This includes adding a page with the sample configuration file, so it
easier to read it when browsing shark's documentation, and a placeholder
for the page that should explain all the available configuration values
that users can tweak.

Since the configuration-related documentation is growing now, we've
moved it to a different folder.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>