New higher-level tests
The new tests include actually running shark against a piece of test
data hosted in a google drive, and then producing all the standard
plots. We are using a nice combination of python2/python3 and matplotlib
(1.3.x in Ubuntu, latest in a virtualenv), so we should be able to pick
errors pretty soon.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>