Adjusting parameters' default values
Now that many runs have gone through, and that we are distributing a
sample configuration file with shark, we offer a better set of default
values. The sample configuration file actually relies on these
parameters to have these default values in order to produce a healthy
run. This is because it does not list every single possible
configuration item (some of them are too advanced to expose them to
users).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>