Organizing code around generate_random_orbits
Firstly, I centralized the code that calculate random points in a sphere
of radius r so it's easy to use it, and so we avoid future mistakes with
coordinate conversion. And secondly, I made the code more compact and
easier to read, making use of the new facilities provided by the
additional methods of the xyz template class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>