New subhalos_view type
This will allow us to implement the all_subhalos() method of the Halo
class in such a way that it will return a subhalos_view object instead
of a new vector, saving multiple copies et al.

Unit tests double-check that the subhalos_view and the underlying
iterator work as intended.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>