Convert mem-per-task to mem-per-cpu
This wasn't a problem in the days when we supported only single-threaded shark
instances, but now that we support multi-threaded instances, and we allow users
to specify a number of CPUs to use per instance via -c, we need to scale the
memory per task/instance so it takes into account that more than one CPU is
used per task.

For example, when using "-V 0 -c 16 -m 10g" (i.e., a single instance to process
subvolume 0 using 16 cores and 10g of memory) we were requesting 160g of memory
(10g per CPU). We now scale this value correctly, and ask the correct amount
(640m per CPU = 10g per task).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>