Improving memory calculations
Now that we were correctly adjusting our memory requests we also needed to
adjust how we used that memory. To correclty do this we need to take into
account the number of CPUs per task, and therefore we needed to swap the two
pieces of logic so #CPUs are calculated first and amount of memory later. After
that, the change itself was simple enough.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>