Building a static library in Windows
Because we haven't declared all the API entry points in our shared
library, MSVC-based compilations doesn't produce a .lib file, and
therefore programs linking to the library fail.

Having a static library solves this issue, but on the other hand we
probably still want to keep using a shared library in systems where we
can link to it (to save space in disk and in memory).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>