numerical_constants fashion emergency
This is a big change, but mostly cosmetic.

First of all there was a big number of constants defined in the
shark::constants namespace that were not used: those are gone.

Secondly, the description of each constant was usually in the same line
as the declaration itself, while now they are properly formatted as
doxygen docs; moreover, each logical group of constants is under a @name
tag as well.

And last but not least, all constants are now defined with type double.
There was no particular reason for using floats in the first place, and
actually a few reasons not to: firstly, outside the constants namespace
we do all (or almost all) our calculations using doubles, so these
constant values were being implicitly promoted to doubles; secondly, the
calculation of the constants themselves was using float numbers with
wide dynamical ranges, and therefore might have created unnecessarily
incorrect constant values (but I didn't double check this), even if they
were being promoted to doubles. And finally, because literal decimal
values in C++ are doubles (unless they have an 'f' suffix) it would have
required a big amount of effort to safely declare those literals as
floats without having warnings in some compilers about literal value
truncations (MSVC was particularly noisy about that).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>