Trying to get rid of truncation warnings in xyz
First of all, the templated ctor was unnecessary, as casting should
happen when translating the arguments, and not when assigning their
values to the inner members. I am offering an non-templated ctor instead
which should take care of the constructions of xyz objects with given
values. On top of that I'm offering a new empty constructor that fills
inner members with zeros.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>