Clarifying how evolution through snapshots works
This change should not affect behavior at all. It simply clarifies both
in the code in the form of comments, and in the logs in the form of more
informative text, how the process of evolving galaxies works for a given
snapshot. Because we evolve using a single snapshot number (say, "i"),
it's easy to think that we evolve AT snapshot "i", while in reality we
evolve FROM snapshot "i" to snapshot "i+1". This means that the main
evolution loop goes through the range [min, max) (i.e., open ended in
max), and that when we write galaxies (after evolution) they logically
are in snapshot "i+1".

In the case of the writing of galaxies, this "i+1" was permeating out
from the main logic loop in shark_runner and into the galaxy_writer
module (and actually only for the HDF5 writer, and not in the ASCII
writer). I've changed this so now the galaxy_writer module doesn't have
to bother with these details.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>