Creating and handing down a _redshift_table
All plots eventually need an instance of the new _redshift_table class
because they always plot quantities at one or more specific redshifts.
This change only adds the new object to the tuple returned by
parse_args, and the reception of the new object in all scripts, but we
still need to start using it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>