common.py's main can print snapshots too
Previously, when calling common.py as a main program we simply printed
the output directory of the plots for a given configuration file. This
was used by our shark-run script, who needed this information to make a
copy of all the plots generated by the plotting scripts, but without
having to re-implement this logic.

Now that we have a redshift table to calculate snapshots, we can now
expose this functionality to external programs that need this
information as well without having to re-implement it. The main use case
for this is again the shark-run script: when plots are requested we should
be calculate in advance which snapshots need to be written by shark, and
modify its configuration file to ensure those snapshots are written to
disk (and therefore the plots succeed). This obviously requires the
knowledge of which redshifts are needed by the plots, which is the only
bit that should be hardcoded (somewhere in the plots centrally,
probably).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>