Using the new redshift table
Different plotting scripts had different ways of specifying snapshots
previously: either they hardcoded the list of snapshots representing a
list of hardcoded redshifts, or they accepted a snapshot as a
command-line argument, in which case it always meant "the snapshot at
redshift 0".  This commit changes both use cases at the same time: in
the first case we change the hardcoded list of snapshots by a hardcoded
list of redshifts, and calculate the snapshots from it. In the second
case we don't use the "snapshot" command-line argument anymore, but
hardcode a redshift 0 in the code, and request the corresponding
snapshot from the redshift table.

Because the "snapshot" "snapshot" positional parameter is now
unnecessary we also remove from the command-line parsing.

Finally, the shark-run script used in HPC systems doesn't need to
specify a hardcoded snapshot anymore.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>