Updating output snapshots list in sample config
The reason for this update is that the actual snapshot representing
redshift 4 is 100 and not 99 (because 100's redshift is closer).

Although this is only a sample configuration file, and shouldn't have
any visible effect on how the code runs, we still want to update it for
two reasons: on the one hand we want to show a nice set of defaults to
users, and a good default list of output snapshots that matches
perfectly the redshifts used by the plots (for the SURFS simulation at
least) is a good thing. On the other hand, if we didn't change this
value here we would have had to update the test script used in Travis
anyway, which uses the sample configuration file, to override this
particular preference and set the correct list of necessary output
snapshots; otherwise the plots generated during the tests run fail
because there is no output for snapshot 100.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>