Removed escape velocity limit for types 2
The escape velocity limits wasn't considering the fact that there is
mass outside the radius at which the galaxy lives, and thus it was been
removed. We also decided (discussing with Danail Obreschkow) that we
want to allow some types 2 to have velocities beyond the escape velocity
as they would anyway have r<1Rvir. 

- Corrected indenting of execution.cpp.