Changing output_snapshots from vector to set
This makes this preference more resilient to duplicate values.

Although this should be a trivial change, it isn't because we were
actually relying on this being a vector, and moreover on its first value
being the highest value of the collection. This is of course not nice,
so not only now we ensure uniqueness of elements, but we also offer a
simpler interface to get the highest element (i.e., the last element of
the set) and to find if a particular snapshot should be outputed or not.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>