Preventing errors when using shark-submit -p
When shark-submit is requested to automatically produce the standard set
of plots, the scripts that produce the plots will fail (or not)
depending on whether shark's configuration file lists the necessary
snapshots that the plots need to read as output snapshots. Instead of
simply relying on users actively adding those snapshots to the
configuration file they use to execute shark, we now make sure that the
snapshots that will be required by the plots are also listed in the
execution.output_snapshots parameter. We do this by asking the python
scripts which snapshots correspond to a particular set of redshifts
(which we know the plots need), and append the resulting list of
snapshots to the aforementioned parameter.

In the future we shouldn't even need to ask *which* redshifts are
needed, because that information is already contained somehow in the
plotting scripts themselves. When that time comes we simply need to ask
the plotting scripts which snapshots it needs, and not which snapshots
correspond to particular redshifts.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>