Improving OpenMP version checking
OpenMP support across the 3.x series of cmake has varied a bit, reflecting the
needs of different compilers. In particular, Intel compilers' support for
OpenMP was not being correctly recognized, as additional linkage needs to
happen to create a fully working OpenMP program, which wasn't being catered for
in our manual test compilation for version checking.

With this commit we are using the newer support for OpenMP version
recognition when available, also adding the required OpenMP libraries to
the linking process of shark to ensure a successful build. When we find
the particular situation of an Intel compiler failing to find OpenMP
support with cmake < 3.9 then we also display a message to the user to
let them know they should try with a newer cmake version.

This should fix #1.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>