Added AGN feedback as an input model in physical_model
both AGN feedback and parameters are now included in the physical model.
In addition the BH accretion rate is given as one of the parameters in
the physical model.

To complete the treatment of QSO feedback I still need to modify the set
of ODEs to compute the additional outflow and ejections rate due to QSO
feedback in starbursts.