ODESolver redesigned for better memory allocation
Originally, ODESolver was designed to evolve an ODE system with a fixed
set of initial conditions through different time steps using a single
object. However, the way PhysicalModel uses the solver is different: it
evolves an ODE system for different sets of initial conditions through
the same period of time. Therefore, the only way to fit the old design
of ODESolver into how PhysicalModel used it was to create a new
ODESolver for each initial condition set (i.e., each galaxy) and evolve
it a single step of delta_t. This situation was obviously not ideal, as
it introduces unnecessary dynamic memory allocation/deallocation for
*each* galaxy that is being evolved.

The new design of ODESolver fits better into the way it ends up being
used. It is initially created with a reference to the ODE system that
will be used throughout but without initial conditions; on the other
hand, the evolve method receives each time the initial conditions, a
delta_t parameter, and the ODE system is reset and then evolved for the
given parameters. This new design then dynamically allocates only one
set of objects for the lifetime of a PhysicalModel, improving the
behavior of shark in multi-threading situations.

Also in this commit we are taking the opportunity to add a deleter class
for the gsl_odeiv2_driver object. This makes the code easier to
maintain, as we now have both a move constructor and an assignment
operator implicitly declared.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>