Avoid accessing MergerTree::halos with operator[]
Using this form we were accidentally creating empty vectors within those
maps for snapshots in which some of the MergerTrees didn't have halos.
This unnecessary memory could actually be quite important: in some
mini-surfs runs with a single subvolume they used ~15% of the total of
allocated memory. This is therefore an important reduction in our memory
usage, which is always welcomed.

To avoid using the operator[] directly on the halos map we now offer a
halos_at(int) method that can be used instead. It checks if there's
something in the halos map for the given snapshot, and if there isn't
then an empty static vector is returned instead.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>