Avoid dynamic memory allocations in PhysicalModel
Each time we evolve a galaxy (and we evolve lots of them) we performed a
memory allocation of a vector of doubles to gather the relevant values
from the galaxy and pass them into the ODE system. This is unnecessary:
we can definitely have a single, long living vector (or even an array)
and use it each time, which is what we're doing now.

In a single-threaded execution this might not have much of an impact,
but in multi-threaded ones this might have been causing contention.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>