Running shark in AppVeyor
We previously didn't run shark in AppVeyor, probably because it was a
bit buggy. After the few fixes we've put in place we can now run it
successfully in Windows when compiled with VS2017, so let's add this
extra step to our AppVeyor pipeline.

On top of running shark, we also run the standard plots upload the
generated PDF files as artifacts of the build for later inspection. This
is something we currently cannot do in TravisCI, so is definitely a
welcomed addition.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>