Portable Halo/Subhalo ID type definitions
Using "long" to store Halos/Subhalo IDs doesn't translate to 64 bits in
all platforms -- most notably in Windows. Thus, a more portable way of
ensuring this is to use the std::int64_t type.

Together with this change of type definition we also need to make sure
we use the correct type when handling these values. The most important
example was when reading/writing from/to hdf5 files, where we use
vectors to transport all IDs between the program and the file. The HDF5
traits we define in traits.h also needed to be adjusted to account for
this change (we had a trait for "long", now we have one for
"std::int64_t" instead).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>