Lognormal sigma should be > 0
After discussing this with Aaron, it would seem that the values of sigma
are constrained to be positive by convention and semantic meaning more
than by explicit mathematical requirements. Taking the absolute value of
whatever we use should then be fine and should yield the same value
distribution.

Also, instead of putting the already-logged numbers, we are making the
log and abs operations more explicit to other readers of the code. This
should also make it easier to change the lognormal parameters in the
future if required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>