Fixed problem with descendant_id
There was a problem when checking whether a galaxy would merge on the
following timestep (with respect to the one being written), as the
comparison between the merging timescale and the deltat of the snapshots
was being made with the wrong set of snapshots.

The calculation of whether galaxies will merge in snapshot+1 was also
moved from galaxy_mergers to galaxy_writer, so that it is defined right
before writing rather than at the beginning of the calculation in
shark_runner.