Attempting to fix descendant_id.
It seems that the previous fixed only deatl with cases of galaxies that
were already type=2 of a central galaxy, but ignored the cases in which
a central may become type 2 in the next snapshot. In that case the
satellites type =2 that central have should point to the central of the
next snapshot rather than their current central (if they will merge on
the next snapshot).