clang-tidy: avoid const rvalues, enable std::move
These were highlighted by the performance-move-const-arg check. Since
the objects being moved were const-qualified, a move couldn't actually
happen, but the code still gave the wrong impression that it actually
did. Changing the variable to be non-const is the correct solution in
this case, since the original intention was always to perform a move.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>