clang-tidy: not moving when not possible/necessary
These were highlighted by the performance-move-const-arg check. These
obey to a different need though: instead of having to correctly declare
a variable to be non-const so we can perform a move on it, now we are
necessarily dealing with variables that *need* to be const, or that are
trivially copyable. In the case of const variables a move cannot be
performed, while in the latter case a move doesn't add any performance
benefit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>