clang-tidy: more override, less virtual
These were highlighted by the modernize-use-override check. On the one
hand using virtual on member function declarations that are not the
initial one is not necessary, while on the other using the override
keyword forces the compiler to fail if the function is not an actual
override of a virtual function. Therefore, using the latter instead of
the former is the best approach.

Using this I actually found that one of the functions that were
overriden in the PhysicalModel class was not declared virtual. However,
since the compiler could see it when compiling the derived class, it
still generated the correct code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>