Moving from value to default initialization
Class members that are themselves of class/struct type do not need to be
explicitly value-initialized with a default member initialization (i.e.,
using the "T t {};" syntax). Instead, we can use default initialization
(i.e., declaring them as "T t;"), which has the same effect of calling
the default constructor on them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>