Cleaning up -Wreturn-type warnings
These functions have been producing warnings for a long time, but we
have kept ignoring them. They are a symptom of a more widespread
problem, which is that a few models (disk size, concentration, etc) are
half-way implemented only, so they cannot be used reliably.

I'm silencing the warnings for the time being, but a longer-term
solution needs to be put in place eventually.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>