Improved OpenMP integer-based functions
Our integer-based "omp_*_for" functions until now required both integer
arguments ("first" and "last") to be of the same integer type. This
requirement makes it a bit difficult to use these functions when the two
values have different data types (e.g., a literal int "0" and a variable
of type std::size_t).

This new flavour of the functions removes the need of having the same
type, and internally computes a common suitable type before turning it
into a signed type for OpenMP usage. We've also given a better name to
this calculated type, calling it omp_int instead of omp_iterator_t.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>