clang-tidy: no more "using namespace std"
These were highlighted by the google-build-using-namespace check. We had
removed this usage from most parts of the code, but there were a few
cases where they still lingered on, specially within the importer
directory that we don't regularly touch.

Some further warnings came out because we import the shark::constants
namespace in bulk in many parts of the code, but a better remedy for
that is probably defining the contents in the shark namespace in the
first place.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>