Using traits to get builtin type names
This removes the need to pass down character literals from different
function calls, which always looked unnecessary and also didn't
centralize the names themselves. As a side effect we'll get nicer error
messages (i.e., more specialized) when an invalid option value is found
in a list-of-X (with X being an integral type).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>