Fixed remaining issue with descendant_id
There was a left over problem that happened very rarely, which consist
of a galaxy being identified as merging on the next snapshot in
merging_galaxies, while having their tmerge recalculated later on in
merging_subhalos, significantly increasing it. Their descendant_id has
to be redefined in the latter case to make sure the galaxy has a
descant_id of -1 before going into galaxy_writer.