Added new component lost_galaxy_mass due to QSO feedback
- added new component to subhalos lost_galaxy_mass which tracks the
outflowing mass due to QSO feedback. In principle this will be treated
as lost mass, but we could in the future experiment with reincorporation
timescales for this reservoir.

- a new component to AllBaryons has been created to track the
lost_galaxy_mass reservoir.

- galaxy mergers and environment have been changed to transfer the
lost_galaxy_mass of satellites to the centrals of their halo.

- physical_model has been extended to include this new component. For
now their rate of change is 0 and hence the previous versions of the
model are fully reproduceable.