Improving documentation for -E option
The documentation (or at least the help message) now contains some
further information about the -E option and the contents of the
parameters file. It also doesn't mention it as "particle position" file,
which points back to the concept of PSO, but rather as a "parameter set"
file which is more generic.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>