Improving internal handling of -E option
This includes making sure the file indicated by the option is actually
there and fail quickly if not, and avoid the calculation of
num_particles when no -E option is given, calculating n_tasks directly
instead. Also the copying of the parameters file still had a hardcoded
path in it and wasn't using the value given by the -E option.

Also in the spirit of making names more generic I've renamed the
variable pointing to the file from posPath to shark_params_file, which
is more aligned with the rest of the variable names.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>