First cut of PSO scripts
This first version is based on the work of Mawson Sammons, who
implemented the initial version of a desktop and HPC-enabled PSO
optimisation technique using the pyswarm package.

Even though we are based on the pyswarm package, we include our own
version of the pso module that comes from that package. This is because
we have a couple of fixes/improvements over it, and the upstream project
seems to be inactive. These fixes include:

 * Fix the maximum number of iterations (there was one too many)
 * Allow users to dump interim results into a file
 * Allow users to perform their own particle evaluation parallelisation

The original code over which these changes are based on can be found
under https://github.com/18838367/Project1819

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>