Correct visibility/accum for evolution_times
The previous declaration of evolution_times made it unreachable from
outside the SharkRunner::impl class, and therefore the implementation of
the operator<< overload for output was not understood by all compilers.
Declaring the structure publicly solves this.

Also, the way we were accumulating times across different threads was
thread-unsafe, so now we are doing it properly: collecting them in
individual per-thread objects and summing them up at the end.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>