Adding documentation checks to Travis
It has happened quite a few times now that the hdf5 output produced by
shark has been enriched, but the online documentation doesn't contain
that information. This check prevents exactly this, forcing users to
update the documentation alongside with additions to the HDF5 output
files.

In order to generate the documentation we need to install the hdf5-tools
package in Ubuntu, which brings the h5dump command line tool. The MacOS
hdf5 brew package already installs this tool.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>