Fixing compilation with OpenMP
The change introduced in 5060a37 had good intentions, but nasty
consequences: it turns out that in a few systems the
OpenMP_CXX_LIBRARIES variable could be empty, and therefore at linking
time the linker needs to be given the OpenMP flag as well. This was
happening implicitly when we were appending the OpenMP flag to the
CMAKE_CXX_FLAGS variable, so when we switched to using
add_compile_options we lost that implicit behavior, and broke
the compilation of OpenMP-enabled shark in certain situations.

The didn't present itself in any of our Travis and AppVeyor instances,
where things compiled correctly. The first platform where we saw this
was our local pleiades cluster, where we are using compilers that are
installed as modules (i.e., outside /usr and /usr/local) which probably
plays a key role on triggering this defect.

This fix doesn't fully revert 5060a37; instead it sticks to using
add_compiler_options and on top of that it adds the OpenMP flag to the
linker flags. This seems to be the accepted solution for adding linker
flags globally. An even better way of doing this of course would be to
use the LINK_FLAGS target property, but we don't use target properties
at all in shark (yet).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>