Appropriate use of Lambda DM from VR catalogs
Shark was allowing the user to use the lambda values of the VR catalogs
if the number of particles in the halo was  larger than some specified
minimum number of particles. The issue was that those values were being
rewritten in the tree_builder as the code was still written for the case
of lambda being taken from a log-normal distribution. This is now fixed.