New execution module under optim
This module contains the code related to actually executing shark,
either normally or in an HPC system. The idea of moving the code to this
module is simply to make the main easier to follow; otherwise nothing
has changed.

To move this code into a new module I had to move also some code to be
under common.py; otherwise I would have had to repeat myself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>