Calculating relative weight
The absolute weight is good, but having a normalized vision of the
constraint weights is even better for users to reason about how they are
used.

This change makes the printing of Constraint objects a bit more
straight-forward, as we probably always want to see the normalized,
relative weight of Constraints.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>