Modified option for ignoring massive root halos
It turns out that the problem of massive halos popping into existence at
low redshift is a common problem in massive halos in velociraptor, and
hence ignoring the whole merger tree in which that halo lives has the
effect of removing a significant fraction of the massive halos.

Hence, I modified the option of ignoring halos that are appearing late
and massive in a way that these halos are not allow to form a galaxy
(hence gas there doesn't cool down).

This is clearly a patchy solution as ideally one would like the halo
catalog to already be pruned before getting to the shark instance, but
in the meantime it does the job.