Avoiding MPL bug triggered by "empty" error bars
There was a bug in matplotlib (versions 1.5.1 and 2.1.x) where error
bars that looked "empty" (in terms of them evaluating to
np.any(error_bars) == False) caused an IndexError to be raised. To avoid
this we simply don't pass down error bars that look like this from our
common.errobars method (e.g., an error bar that looks like [0, 0], which
would produce no visible output anyway), and make sure we use this
method instead of the Axes.errorbar method directly where this has been
an issue.

The version of matplotlib shipped with Ubuntu Xenial was exactly 1.5.1,
which is why we were suffering from this bug when updating to the newer
Ubuntu images in Travis.

See https://github.com/matplotlib/matplotlib/issues/9699 for details.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>