Porting code to use new headers
The rule of thumb is that header files defining classes and
functions that take references or pointers to one of Subhalo, Halo,
Galaxy, MergerTree or TotalBaryon keep including the components.h
header, which forward-defines these classes and their *Ptr counterparts.
On the other hand .cpp files that actually need the full definition
include the individual headers (e.g., halo.h) to get access to the full
definition.

This effectively makes our code faster to re-compile when developing and
making changes to the different structure-related classes like Halo and
Galaxy.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>