New type2_galaxies_view class
This new class will allow us to directly iterate over a vector of Galaxy
objects whose type is TYPE2 but without having to create a new vector
with copies of the original Galaxy objects, and therefore enabling us to
access them by reference, and modify them as necessary. Because the
class offers being() and end() methods it means it can be used in
range-for loops, algorithms and more.

Internally, the type2_galaxies_view class defines its own iterator class
where the actual logic happens. The iterator takes the begin() and end()
of a vector of Galaxy objects. Both at creation time and after normal
increments (i.e., operator++ is called), the iterator continues
advancing through the original vector of Galaxy objects until a TYPE2
galaxy is found.

This new class will be vital when switching the Subhalos.galaxy member
from being a vector of GalaxyPtr to a vector of Galaxy objects.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>